requirements.txt
fastapi==0.108.0
passlib==1.7.4
pydantic==1.10.13
python_jose==3.3.0
SQLAlchemy==1.4.46
psycopg==3.1.7
psycopg2-binary==2.9.9
uvicorn==0.17.6
email-validator==2.1.0.post1
python-multipart==0.0.5
alembic==1.13.1
python-dotenv==0.21.0